requirements.txt
Flask
Tinydb
mtranslate
python-telegram-bot==13.15
requests
openai